Refactor ms resources
feat(ms-resources): inventario granular por item de catalogo y categorias dinamicas

Refactor del modelo de recursos: el inventario deja de gestionarse por
categoria (enum fijo) y pasa a gestionarse por item de catalogo, con una
entidad Categoria dinamica identificada por UUID.

- Modelo: nueva entidad Categoria (tabla categorias, codigo unico + semilla
de 6 categorias). catalogo_items e inventario referencian UUID
(categoria_id, item_catalogo_id) en lugar del enum.
- Inventario por item_catalogo_id; stock y umbrales pasan a BIGINT (long).
- movimientos_inventario registra item_catalogo_id y cantidades en BIGINT.
- Migraciones Flyway V8-V12 con migracion de datos que preserva inventario y
movimientos existentes, y limpieza de columnas legacy.
- Endpoints nuevos: POST /catalogo/items (CATALOGO_GESTIONAR) que inicializa
filas de inventario en centros, CRUD de categorias (CATEGORIA_GESTIONAR) y
GET /centros/{id}/inventario/resumen-categorias. El listado de catalogo
filtra por categoriaId.
- RBAC: nuevos permisos CATEGORIA_GESTIONAR y CATALOGO_GESTIONAR para
ADMINISTRADOR.
- Excepciones nuevas (CategoriaNoEncontrada, CategoriaDuplicada,
ItemCatalogoNoEncontrado) integradas al manejador global RFC 7807.
- DTOs: + CategoriaResponse, InventarioItemResponse,
ResumenInventarioCategoriaResponse, CrearItemCatalogoRequest,
Crear/ActualizarCategoriaRequest; eliminado InventarioCategoriaResponse.
Tests unitarios actualizados y nuevos tests de integracion de inventario.
- Se incluye .gitignore correspondiente al proyecto.

+++ src/main/java/cl/catastrofescl/resources/exception/StockInsuficienteException.java
@@ -2,7 +2,7 @@ package cl.catastrofescl.resources.exception;
 
 public class StockInsuficienteException extends RuntimeException {
 
-    public StockInsuficienteException(int stockActual, int cantidadSolicitada) {
+    public StockInsuficienteException(long stockActual, long cantidadSolicitada) {
         super("Stock insuficiente: actual=" + stockActual + ", solicitado=" + cantidadSolicitada);
     }
 }
